+++ R/build-home-authors.R
@@ -1,22 +1,33 @@
-data_authors <- function(pkg = ".") {
+data_authors <- function(pkg = ".", roles = default_roles()) {
   pkg <- as_pkgdown(pkg)
   author_info <- data_author_info(pkg)
 
   all <- pkg %>%
     pkg_authors() %>%
-    purrr::map(author_list, author_info)
+    purrr::map(author_list, author_info, pkg = pkg)
 
   main <- pkg %>%
-    pkg_authors(c("aut", "cre", "fnd")) %>%
-    purrr::map(author_list, author_info)
+    pkg_authors(roles) %>%
+    purrr::map(author_list, author_info, pkg = pkg)
 
-  needs_page <- length(main) != length(all)
+  more_authors <- length(main) != length(all)
+
+  comments <- pkg %>%
+    pkg_authors() %>%
+    purrr::map(author_list, author_info, pkg = pkg) %>%
+    purrr::map("comment") %>%
+    purrr::compact() %>%
+    length() > 0
 
   print_yaml(list(
     all = all,
     main = main,
-    needs_page = needs_page
+    needs_page = more_authors || comments
   ))
+}
+
+default_roles <- function() {
+  c("aut", "cre", "fnd")
 }
 
 pkg_authors <- function(pkg, role = NULL) {
@@ -59,28 +70,53 @@ data_author_info <- function(pkg = ".") {
 
 data_home_sidebar_authors <- function(pkg = ".") {
   pkg <- as_pkgdown(pkg)
-  data <- data_authors(pkg)
+roles <- pkg$meta$authors$sidebar$roles %||% default_roles()
+data <- data_authors(pkg, roles)
 
-  authors <- data$main %>% purrr::map_chr(author_desc)
+  authors <- data$main %>% purrr::map_chr(author_desc, comment = FALSE)
+
+  bullets <- c(
+      markdown_inline(
+        pkg$meta$authors$sidebar$before,
+        pkg = pkg,
+        where = c("authors", "sidebar", "before")
+      ),
+      authors,
+      markdown_inline(
+        pkg$meta$authors$sidebar$after,
+        pkg = pkg,
+        where = c("authors", "sidebar", "after")
+      )
+  )
+
   if (data$needs_page) {
-    authors <- c(authors, "<a href='authors.html'>All authors...</li>")
+    bullets <- c(bullets, "<a href='authors.html'>More on authors...</li>")
   }
 
-  sidebar_section("Developers", authors)
+  sidebar_section("Developers", bullets)
 }
 
 build_authors <- function(pkg = ".") {
   pkg <- as_pkgdown(pkg)
 
-  data <- list(
-    pagetitle = "Authors",
-    authors = data_authors(pkg)$all
-  )
+  data <- data_authors_page(pkg)
 
   render_page(pkg, "authors", data, "authors.html")
 }
 
-author_name <- function(x, authors) {
+data_authors_page <- function(pkg) {
+   data <- list(
+    pagetitle = "Authors",
+    authors = data_authors(pkg)$all
+  )
+
+  data$before <- markdown_block(pkg$meta$authors$before, pkg = pkg)
+  data$after <- markdown_block(pkg$meta$authors$after, pkg = pkg)
+
+  return(data)
+}
+
+author_name <- function(x, authors, pkg) {
   name <- format_author_name(x$given, x$family)
 
   if (!(name %in% names(authors)))
@@ -89,7 +125,11 @@ author_name <- function(x, authors) {
   author <- authors[[name]]
 
   if (!is.null(author$html)) {
-    name <- author$html
+    name <- markdown_inline(
+      author$html,
+      pkg = pkg,
+      where = c("authors", name, "html")
+    )
   }
 
   if (is.null(author$href)) {
@@ -109,8 +149,8 @@ format_author_name <- function(given, family) {
   }
 }
 
-author_list <- function(x, authors_info, comment = FALSE) {
-  name <- author_name(x, authors_info)
+author_list <- function(x, authors_info = NULL, comment = FALSE, pkg) {
+  name <- author_name(x, authors_info, pkg = pkg)
 
   roles <- paste0(role_lookup[x$role], collapse = ", ")
   substr(roles, 1, 1) <- toupper(substr(roles, 1, 1))