+++ R/build-home-index.R
@@ -15,7 +15,7 @@ build_home_index <- function(pkg = ".", quiet = TRUE) {
     data$index <- linkify(pkg$desc$get("Description")[[1]])
   } else {
     local_options_link(pkg, depth = 0L)
-    data$index <- markdown(src_path)
+    data$index <- markdown(src_path, pkg = pkg)
   }
   render_page(pkg, "home", data, "index.html", quiet = quiet)
 
@@ -75,12 +75,13 @@ data_home_sidebar <- function(pkg = ".") {
     community = data_home_sidebar_community(pkg),
     citation = data_home_sidebar_citation(pkg),
     authors = data_home_sidebar_authors(pkg),
-    dev = sidebar_section("Dev Status", "placeholder")
+    dev = sidebar_section("Dev Status", "placeholder"),
+    toc = data_home_toc(pkg)
   )
 
   if (is.null(pkg$meta$home$sidebar$structure)) {
     sidebar_html <- paste0(
-      purrr::compact(sidebar_components),
+      purrr::compact(sidebar_components[default_sidebar_structure()]),
       collapse = "\n"
     )
     return(sidebar_html)
@@ -100,26 +101,12 @@ data_home_sidebar <- function(pkg = ".") {
       set_names(names(components))
   )
 
-  missing <- setdiff(sidebar_structure, names(sidebar_components))
-
-  if (length(missing) > 0) {
-    missing_components <- lapply(
-      missing, append,
-      c("home", "sidebar", "components"),
-      after = 0
-    )
-    missing_fields <- pkgdown_fields(pkg = pkg, fields = missing_components)
-
-    abort(
-      sprintf(
-        "Can't find component%s %s.",
-        if (length(missing) > 1) "s" else "",
-        paste0(
-          missing_fields, collapse = " nor "
-        )
-      )
-    )
-  }
+  check_components(
+    needed = sidebar_structure,
+    present = names(sidebar_components),
+    where = c("home", "sidebar", "components"),
+    pkg = pkg
+  )
 
   sidebar_final_components <- purrr::compact(
     sidebar_components[sidebar_structure]
@@ -135,20 +122,17 @@ default_sidebar_structure <- function() {
 
 data_home_component <- function(component, component_name, pkg) {
 
-  if (!all(c("title", "html") %in% names(component))) {
-    abort(
-      sprintf(
-        "Can't find %s for the component %s",
-        paste0(
-          c("title", "html")[!c("title", "html") %in% names(component)],
-          collapse = " nor "
-          ),
-        pkgdown_field(pkg = pkg, "home", "sidebar", "components", component_name)
-        )
-      )
-  }
+  check_components(
+    needed = c("title", "text"),
+    present = names(component),
+    where = c("home", "sidebar", "components", component_name),
+    pkg = pkg
+  )
 
-  sidebar_section(component$title, bullets = component$html)
+  sidebar_section(
+    component$title,
+    bullets = markdown_block(component$text, pkg = pkg)
+  )
 }
 
 data_home_sidebar_links <- function(pkg = ".") {
@@ -168,13 +152,20 @@ data_home_sidebar_links <- function(pkg = ".") {
   sidebar_section("Links", links)
 }
 
+data_home_toc <- function(pkg) {
+  sidebar_section(
+    "Table of contents",
+    '<nav id="toc" data-toggle="toc" class="sticky-top"></nav>'
+  )
+}
+
 sidebar_section <- function(heading, bullets, class = make_slug(heading)) {
   if (length(bullets) == 0)
     return(character())
 
   paste0(
     "<div class='", class, "'>\n",
-    "<h2>", heading, "</h2>\n",
+    "<h2 data-toc-skip>", heading, "</h2>\n",
     "<ul class='list-unstyled'>\n",
     paste0("<li>", bullets, "</li>\n", collapse = ""),
     "</ul>\n",