+++ R/build-logo.R
@@ -6,23 +6,3 @@ build_logo <- function(pkg = ".") {
     file_copy_to(pkg, logo_path, from_dir = path_dir(logo_path))
   }
 }
-
-find_logo <- function(path) {
-  logo_path <- path(path, "logo.svg")
-  if (file_exists(logo_path))
-    return(logo_path)
-
-  logo_path <- path(path, "man", "figures", "logo.svg")
-  if (file_exists(logo_path))
-    return(logo_path)
-
-  logo_path <- path(path, "logo.png")
-  if (file_exists(logo_path))
-    return(logo_path)
-
-  logo_path <- path(path, "man", "figures", "logo.png")
-  if (file_exists(logo_path))
-    return(logo_path)
-
-  NULL
-}