+++ R/markdown.R
@@ -1,9 +1,9 @@
-markdown <- function(path = NULL, ..., strip_header = FALSE) {
+markdown <- function(path = NULL, ..., strip_header = FALSE, pkg) {
   tmp <- tempfile(fileext = ".html")
   on.exit(unlink(tmp), add = TRUE)
 
   if (rmarkdown::pandoc_available("2.0")) {
-    from <- "markdown+gfm_auto_identifiers-citations"
+    from <- "markdown+gfm_auto_identifiers-citations+emoji"
   } else if (rmarkdown::pandoc_available("1.12.3")) {
     from <- "markdown_github-hard_line_breaks+tex_math_dollars+tex_math_single_backslash+header_attributes"
   } else {
@@ -39,6 +39,7 @@ markdown <- function(path = NULL, ..., strip_header = FALSE) {
 
   downlit::downlit_html_node(xml)
   tweak_md_links(xml)
+  tweak_all_links(xml, pkg = pkg)
   tweak_anchors(xml, only_contents = FALSE)
 
   # Extract body of html - as.character renders as xml which adds
@@ -57,7 +58,7 @@ markdown <- function(path = NULL, ..., strip_header = FALSE) {
   )
 }
 
-markdown_text <- function(text, ...) {
+markdown_text <- function(text, pkg = pkg, ...) {
   if (identical(text, NA_character_) || is.null(text)) {
     return(NULL)
   }
@@ -66,5 +67,49 @@ markdown_text <- function(text, ...) {
   on.exit(unlink(tmp), add = TRUE)
 
   write_lines(text, path = tmp)
-  markdown(tmp, ...)
+  markdown(tmp, ..., pkg = pkg)
+}
+
+
+markdown_text_children <- function(text, pkg, ...) {
+
+  html <- markdown_text(text, pkg = pkg, ...)
+  html %>%
+    xml2::read_html() %>%
+    xml2::xml_child() %>% # body
+    xml2::xml_children()
+
+}
+
+markdown_inline <- function(text, pkg, where, ...) {
+
+  if (is.null(text)) {
+    return(NULL)
+  }
+
+  children <- markdown_text_children(text, pkg = pkg, ...)
+
+  if (length(children) > 1) {
+    abort(
+      sprintf(
+        "Can't use a block element here, need an inline element: \n %s \n%s",
+        what = pkgdown_field(pkg = pkg, where),
+        text
+      )
+    )
+  }
+
+  paste0(xml2::xml_contents(children), collapse="")
+
+}
+
+markdown_block <- function(text, pkg, ...) {
+
+  if (is.null(text)) {
+    return(NULL)
+  }
+
+  children <- markdown_text_children(text, pkg = pkg, ...)
+  output <- paste0(as.character(children), collapse="")
+  gsub("\\\n", "", output)
 }