+++ R/navbar.R
@@ -11,13 +11,11 @@ data_navbar <- function(pkg = ".", depth = 0L) {
   components[names(components_meta)] <- components_meta
   components <- purrr::compact(components)
 
-  # Any unplaced components go to the right of the left navbar
   right_comp <- intersect(structure$right, names(components))
   left_comp <- intersect(structure$left, names(components))
-  extra_comp <- setdiff(names(components), c(left_comp, right_comp))
 
   # Backward compatiblity
-  left <- navbar$left %||% components[c(left_comp, extra_comp)]
+  left <- navbar$left %||% components[left_comp]
   right <- navbar$right %||% components[right_comp]
 
   list(
@@ -70,9 +68,12 @@ navbar_components <- function(pkg = ".") {
   }
   menu$news <- navbar_news(pkg)
 
-  if (!is.null(pkg$github_url)) {
-    menu$github <- menu_icon("github", pkg$github_url, style = "fab")
-  }
+  menu$github <- switch(
+    repo_type(pkg),
+    github = menu_icon("github", repo_home(pkg), style = "fab"),
+    gitlab = menu_icon("gitlab", repo_home(pkg), style = "fab"),
+    NULL
+  )
 
   menu <- c(menu, navbar_articles(pkg))
 
@@ -92,9 +93,6 @@ navbar_articles <- function(pkg = ".") {
     menu$intro <- menu_link("Get started", intro$file_out)
   }
 
-  if (!is.null(pkg$repo$url$home)) {
-    menu$github <- menu_icon("github", repo_home(pkg), style = "fab")
-  }
 
   meta <- pkg$meta
   if (!has_name(meta, "articles")) {
@@ -155,17 +153,15 @@ menu_spacer <- function() {
 # Testing helpers ---------------------------------------------------------
 # Simulate minimal package structure so we can more easily test
 
-pkg_navbar <- function(
-                           meta = NULL,
-                           vignettes = pkg_navbar_vignettes(),
-                           github_url = NULL) {
+pkg_navbar <- function(meta = NULL, vignettes = pkg_navbar_vignettes(),
+                       github_url = NULL) {
   structure(
     list(
       package = "test",
       src_path = file_temp(),
       meta = meta,
       vignettes = vignettes,
-      github_url = github_url
+      repo = list(url = list(home = github_url))
     ),
     class = "pkgdown"
   )