+++ R/repo.R
@@ -1,3 +1,15 @@
+repo_type <- function(pkg) {
+  home <- repo_home(pkg) %||% ""
+
+  if (grepl("^https?://github\\..+/", home)) {
+    "github"
+  } else if (grepl("^https?://gitlab\\..+/", home)) {
+    "gitlab"
+  } else {
+    "other"
+  }
+}
+
 repo_home <- function(pkg, paths) {
   pkg$repo$url$home
 }
@@ -32,6 +44,12 @@ repo_auto_link <- function(pkg, text) {
   if (!is.null(url$issue)) {
     issue_link <- paste0("<a href='", url$issue, "\\1'>#\\1</a>")
     text <- gsub("#(\\d+)", issue_link, text, perl = TRUE)
+
+    if (!is.null(pkg$repo$jira_projects)) {
+      issue_link <- paste0("<a href='", url$issue, "\\1\\2'>\\1\\2</a>")
+      issue_regex <- paste0("(", paste0(pkg$repo$jira_projects, collapse = "|"),")(-\\d+)")
+      text <- gsub(issue_regex, issue_link, text, perl = TRUE)
+    }
   }
 
   text