+++ tests/testthat/test-navbar.R
@@ -1,9 +1,12 @@
-test_that("adds github link when available", {
+test_that("adds github/gitlab link when available", {
   verify_output(test_path("test-navbar/github.txt"), {
     pkg <- pkg_navbar()
     navbar_components(pkg)
 
-    pkg <- pkg_navbar(github_url = "https://github.org/r-lib/pkgdown")
+    pkg <- pkg_navbar(github_url = "https://github.com/r-lib/pkgdown")
+    navbar_components(pkg)
+
+    pkg <- pkg_navbar(github_url = "https://gitlab.com/r-lib/pkgdown")
     navbar_components(pkg)
   })
 })
@@ -69,3 +72,20 @@ test_that("can control articles navbar through articles meta", {
 
 })
 
+test_that("data_navbar() works by default", {
+  pkg <- as_pkgdown(test_path("assets/news-multi-page"))
+  expect_snapshot(data_navbar(pkg))
+})
+
+test_that("data_navbar() can re-order default elements", {
+  pkg <- as_pkgdown(test_path("assets/news-multi-page"))
+  pkg$meta$navbar$structure$right <- c("news")
+  pkg$meta$navbar$structure$left <- c("github", "reference")
+  expect_snapshot(data_navbar(pkg))
+})
+
+test_that("data_navbar()can remove elements", {
+  pkg <- as_pkgdown(test_path("assets/news-multi-page"))
+  pkg$meta$navbar$structure$left <- c("github")
+  expect_snapshot(data_navbar(pkg))
+})