+++ tests/testthat/test-repo.R
@@ -17,6 +17,15 @@ test_that("issues are automatically linked", {
   expect_equal(repo_auto_link(pkg, "#123"), "<a href='TEST/123'>#123</a>")
 })
 
+test_that("Jira issues are automatically linked", {
+  pkg <- list(repo = repo_meta(issue = "TEST/"))
+  pkg$repo$jira_projects <- c("JIRA", "OTHER")
+  expect_equal(repo_auto_link(pkg, "JIRA-123"), "<a href='TEST/JIRA-123'>JIRA-123</a>")
+  expect_equal(repo_auto_link(pkg, "OTHER-4321"), "<a href='TEST/OTHER-4321'>OTHER-4321</a>")
+  # but only the jira projects specified are caught
+  expect_equal(repo_auto_link(pkg, "NOPE-123"), "NOPE-123")
+})
+
 # repo_source -------------------------------------------------------------
 
 test_that("repo_source() truncates automatically", {
@@ -90,4 +99,16 @@ test_that("meta overrides autodetection", {
 test_that("returns NULL if no urls found", {
   desc <- desc::desc(text = "URL: https://pkgdown.r-lib.org")
   expect_equal(package_repo(desc, list()), NULL)
+})
+
+test_that("repo_type detects repo type", {
+  repo_type2 <- function(url) {
+   repo_type(list(repo = list(url = list(home = url))))
+  }
+
+  expect_equal(repo_type2("https://github.com/r-lib/pkgdown"), "github")
+  expect_equal(repo_type2("https://github.r-lib.com/pkgdown"), "github")
+  expect_equal(repo_type2("https://gitlab.com/r-lib/pkgdown"), "gitlab")
+  expect_equal(repo_type2("https://gitlab.r-lib.com/pkgdown"), "gitlab")
+  expect_equal(repo_type2(NULL), "other")
 })